refactor file generator to own module
